requirements.txt
pulp
openai
python-dotenv
pandas
